@@ -7,10 +7,10 @@ namespace MinVerTests.Lib.Infra
 {
     internal sealed class TestLogger : ILogger
     {
-#if NET6_0_OR_GREATER
-        private readonly List<LogMessage> messages = new();
+#if NET8_0_OR_GREATER
+        private readonly List<LogMessage> messages = [];
 #else
-        private readonly List<LogMessage> messages = new List<LogMessage>();
+        private readonly List<LogMessage> messages = new();
 #endif
 
         public bool IsTraceEnabled => true;
