@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
@@ -70,7 +69,7 @@ git tag 1.1.0 -a -m '.'
 
             await EnsureEmptyRepositoryAndCommit(path);
 
-            foreach (var command in historicalCommands.Split(new[] { '\r', '\n', }, StringSplitOptions.RemoveEmptyEntries))
+            foreach (var command in historicalCommands.ToNonEmptyLines())
             {
                 var nameAndArgs = command.Split(" ", 2);
                 _ = await ReadAsync(nameAndArgs[0], nameAndArgs[1], path);
