@@ -19,7 +19,7 @@ public static class CommitAfterPreReleaseTag
         await Git.Tag(path, "2.3.4-alpha.5");
         await Git.Commit(path);
 
-        var expected = Package.WithVersion(2, 3, 4, new[] { "alpha", "5", }, 1);
+        var expected = Package.WithVersion(2, 3, 4, ["alpha", "5",], 1);
 
         // act
         var (actual, _, _) = await Sdk.BuildProject(path);
