@@ -21,7 +21,7 @@ public static class MinimumMajorMinorAfterTag
 
         var envVars = ("MinVerMinimumMajorMinor".ToAltCase(), "3.0");
 
-        var expected = Package.WithVersion(3, 0, 0, new[] { "alpha", "0", }, 1);
+        var expected = Package.WithVersion(3, 0, 0, ["alpha", "0",], 1);
 
         // act
         var (actual, _, _) = await Sdk.BuildProject(path, envVars: envVars);
