@@ -14,7 +14,7 @@ public static class NoRepo
         // arrange
         var path = MethodBase.GetCurrentMethod().GetTestDirectory();
         await Sdk.CreateProject(path);
-        var expected = Package.WithVersion(0, 0, 0, new[] { "alpha", "0", });
+        var expected = Package.WithVersion(0, 0, 0, ["alpha", "0",]);
 
         // act
         var (actual, sdkStandardOutput, _) = await Sdk.BuildProject(path);
