@@ -24,7 +24,7 @@ public static class OptionMasking
         var envVars = ("MinVerAutoIncrement".ToAltCase(), "minor");
         var args = $"--auto-increment {value}";
 
-        var expected = Package.WithVersion(2, 3, 5, new[] { "alpha", "0", }, 1);
+        var expected = Package.WithVersion(2, 3, 5, ["alpha", "0",], 1);
 
         // act
         var (cliStandardOutput, _) = await MinVerCli.ReadAsync(path, args: args, envVars: envVars);
@@ -45,7 +45,7 @@ public static class OptionMasking
         var envVars = ("MinVerBuildMetadata", "build.123");
         var args = $"--build-metadata {value}";
 
-        var expected = Package.WithVersion(0, 0, 0, new[] { "alpha", "0", });
+        var expected = Package.WithVersion(0, 0, 0, ["alpha", "0",]);
 
         // act
         var (cliStandardOutput, _) = await MinVerCli.ReadAsync(path, args: args, envVars: envVars);
@@ -70,7 +70,7 @@ public static class OptionMasking
         var envVars = ("MinVerDefaultPreReleaseIdentifiers".ToAltCase(), "preview.0");
         var args = $"--default-pre-release-identifiers {value}";
 
-        var expected = Package.WithVersion(2, 3, 5, new[] { "alpha", "0", }, 1);
+        var expected = Package.WithVersion(2, 3, 5, ["alpha", "0",], 1);
 
         // act
         var (cliStandardOutput, _) = await MinVerCli.ReadAsync(path, args: args, envVars: envVars);
@@ -95,7 +95,7 @@ public static class OptionMasking
         var envVars = ("MinVerDefaultPreReleasePhase".ToAltCase(), "preview");
         var args = $"--default-pre-release-phase {value}";
 
-        var expected = Package.WithVersion(2, 3, 5, new[] { "alpha", "0", }, 1);
+        var expected = Package.WithVersion(2, 3, 5, ["alpha", "0",], 1);
 
         // act
         var (cliStandardOutput, _) = await MinVerCli.ReadAsync(path, args: args, envVars: envVars);
@@ -143,7 +143,7 @@ public static class OptionMasking
         var envVars = ("MinVerTagPrefix", "v.");
         var args = $"--tag-prefix {value}";
 
-        var expected = Package.WithVersion(2, 3, 4, new[] { "alpha", "5", });
+        var expected = Package.WithVersion(2, 3, 4, ["alpha", "5",]);
 
         // act
         var (cliStandardOutput, _) = await MinVerCli.ReadAsync(path, args: args, envVars: envVars);
