@@ -18,7 +18,7 @@ public static class PreReleaseTag
         await Git.Commit(path);
         await Git.Tag(path, "2.3.4-alpha.5");
 
-        var expected = Package.WithVersion(2, 3, 4, new[] { "alpha", "5", });
+        var expected = Package.WithVersion(2, 3, 4, ["alpha", "5",]);
 
         // act
         var (actual, _, _) = await Sdk.BuildProject(path);
