@@ -20,7 +20,7 @@ public static class TwoCommitsAfterPreReleaseTag
         await Git.Commit(path);
         await Git.Commit(path);
 
-        var expected = Package.WithVersion(2, 3, 4, new[] { "alpha", "5", }, 2);
+        var expected = Package.WithVersion(2, 3, 4, ["alpha", "5",], 2);
 
         // act
         var (actual, _, _) = await Sdk.BuildProject(path);
